admin subcategory manager

+++ admin/js/wallpaper.js
@@ -3,10 +3,10 @@
 SQL='SELECT * FROM `wallpaper_wallpaper`';
 popupData={};
 categoryManager={};
+subcategoryManager={};
 
 // ===== ACTIONS =====
 сategorySelectGetData();
-getGaleryData(SQL);
 
 
 // ===== LISTENERS =====
@@ -51,19 +51,88 @@ $('body').delegate('button[name="category-manager_delete"]','click',function(){
    if (arr.length>0) {
     alert('Объект содержит вложенные элементы. Удаление невозможно!');
   }else{
-    var ask=confirm('Категория '+categoryManager.category+' будет удалена!');
-    if (ask) {
-      var sql='DELETE FROM `wallpaper_category` WHERE `category`="'+
-      categoryManager.category+'"';
-      $.post('php/update.php',{data:sql}, сategorySelectGetData);
-    }
+    var sql='SELECT * FROM `wallpaper_wallpaper` WHERE `category`="'+
+    categoryManager.category+'"';
+    $.post('../wallpaper/php/select.php',{data:sql},function (data) {
+      var arr=JSON.parse(data);
+      if (arr.length>0) {
+        alert('Объект содержит вложенные элементы. Удаление невозможно!');
+      }else{
+        var ask=confirm('Категория '+categoryManager.category+' будет удалена!');
+        if (ask) {
+          var sql='DELETE FROM `wallpaper_category` WHERE `category`="'+
+          categoryManager.category+'"';
+          $.post('php/update.php',{data:sql}, сategorySelectGetData);
+        }
+      }
+    });
   }});
 });
+$('select[name="subcategory-manager_select"]').change(subCategoryManagerGetData);
+$('body').delegate('button[name="subcategory-manager_edit"]','click',function(){
+  $('input[name="subcategory-manager_input"]').prop('disabled',true);
+  $(this).parent().siblings().children('input').prop('disabled',false);
+});
+$('body').delegate('button[name="subcategory-manager_save"]','click',function(){
+  var subcategory=$(this).parent().siblings().children('input').prop('value').trim();
+  subcategoryManager.subcategory=this.value;
+  sql='UPDATE `wallpaper_subcategory` SET `subcategory`="'+
+  subcategory+'" WHERE `subcategory`="'+subcategoryManager.subcategory+'"';
+  $.post('php/update.php',{data:sql});
+  sql='UPDATE `wallpaper_wallpaper` SET `subcategory`="'+
+  subcategory+'" WHERE `subcategory`="'+subcategoryManager.subcategory+'"';
+  $.post('php/update.php',{data:sql},сategorySelectGetData);
+});
+$('body').delegate('button[name="subcategory-manager_delete"]','click',function(){
+  subcategoryManager.subcategory=this.value;
+  var sql=
+  'SELECT * FROM `wallpaper_wallpaper` WHERE `subcategory`="'+this.value+'"';
+  $.post('../wallpaper/php/select.php',{data:sql},function (data){
+    var arr=JSON.parse(data);
+    if (arr.length>0) {
+      alert('Объект содержит вложенные элементы. Удаление невозможно!');
+    }else {
+      var ask=confirm('Подкатегория '+subcategoryManager.subcategory+' будет удалена!');
+      if (ask) {
+        var sql='DELETE FROM `wallpaper_subcategory` WHERE `subcategory`="'+
+        subcategoryManager.subcategory+'"';
+        $.post('php/update.php',{data:sql}, subCategoryManagerGetData);
+      }
+    }
+  });
+});
+$('.subcategory-manager_add').click(function () {
+  var category=$('select[name="subcategory-manager_select"]').prop('value').trim();
+  var subcategory=prompt('Название подкатегории:');
+  if(subcategory!=''&&subcategory!=null){
+    subcategory=subcategory.trim();
+    var sql=
+    'INSERT INTO `wallpaper_subcategory`(`category`,`subcategory`)'+
+    'VALUES("'+category+'","'+subcategory+'")';
+    $.post('php/update.php',{data:sql}, subCategoryManagerGetData());
+  }
+});
+$('select[name="admin-category_filter"]').change(categorySelectGalery);
+$('select[name="admin-subcategory_filter"]').change(filterGalery);
+$('button[name="search-button"]').click(function () {
+ var sql='SELECT * FROM `wallpaper_wallpaper`';
+ var wallpaper=$('input[name="search-input"]').prop('value').trim();
+ if (wallpaper!=''){
+  sql='SELECT * FROM `wallpaper_wallpaper`'+
+  ' WHERE `wallpaper` LIKE "%'+wallpaper+'%"';}
+  getGaleryData(sql);
+});
+$('button[name="search-resset"]').click(function () {
+  $('input[name="search-input"]').prop('value','');
+  getGaleryData(SQL);
+});
+
 
 // ===== FUNCTIONS =====
 function сategorySelectGetData () {
  var sql='SELECT * FROM `wallpaper_category`';
  $.post('../wallpaper/php/select.php',{data:sql}, categorySelectForm);
+ getGaleryData(SQL);
 }
 function categorySelectForm (data) {
   var arr=JSON.parse(data);
@@ -74,8 +143,11 @@ function categorySelectForm (data) {
  }
  $('select[name="wallpaperCategory"]').html(list);
  $('select[name="editPopupCategory"]').html(list);
+ $('select[name="subcategory-manager_select"]').html(list);
+ categoryFilterGetData();
  subCategorySelectGetData();
- categoryManagerList(arr);
+ categoryManagerForm(arr);
+ subCategoryManagerGetData();
 }
 function subCategorySelectGetData () {
   var category=
@@ -92,6 +164,7 @@ function subCategorySelectForm (data) {
    arr[i].subcategory+'</option>';
  }
  $('select[name="wallpaperSubCategory"]').html(list);
+ subCategoryManagerGetData(arr);
 }
 function wallPaperImageUpload () {
   var image=$('input[name="wallpaperImage"]')[0];
@@ -145,7 +218,7 @@ function useGaleryData (data) {
 }
 function galeryItemClone (arr) {
   $('.wallpaper-galery_item:gt(0)').remove();
-  for (var i = 0; i < arr.length-1; i++) {
+  for (var i = 1; i < arr.length; i++) {
     var node=document.getElementsByClassName('wallpaper-galery_item')[0];
     var cloneNode=node.cloneNode(true);
     var galery=document.getElementsByClassName('wallpaper-admin_galery')[0];
@@ -155,6 +228,8 @@ function galeryItemClone (arr) {
 }
 function fillGaleryData (arr) {
   var itemArr=$('.wallpaper-galery_item');
+  if (arr.length<1){$(itemArr).css({'display':'none'});}
+  else {$(itemArr).css({'display':'block'});}
   var infoArr=$('.wallpaper-gallery_info');
   var imageArr=$('.wallpaper-galery_image');
   var interiorArr=$('.wallpaper-galery_interior');
@@ -218,8 +293,7 @@ function saveEditPopUpData () {
   $.post('php/update.php',{data:sql}, function () {
     getGaleryData(SQL)
   });
-  $('.edit-popup_wrapper').fadeOut();
-}
+  $('.edit-popup_wrapper').fadeOut();}
 }
 function itemGaleryDelete () {
   var ask=confirm('Элемент будет удален!');
@@ -239,7 +313,7 @@ function itemGaleryDelete () {
     });
   }
 }
-function categoryManagerList (arr) {
+function categoryManagerForm (arr) {
   $('.category-manager_item:gt(0)').remove();
   for (var i = 1; i < arr.length; i++) {
     var tr=$('.category-manager_item');
@@ -247,9 +321,9 @@ function categoryManagerList (arr) {
     var table=$('.category-manager_table');
     table[0].appendChild(cloneTr);
   }
-  categoryManagerListData(arr);
+  categoryManagerFormData(arr);
 }
-function categoryManagerListData (arr) {
+function categoryManagerFormData (arr) {
   var idArr=$('.category-manager_id');
   var categoryArr=$('input[name="category-manager_input"]'); 
   var saveArr=$('button[name="category-manager_save"]');
@@ -273,4 +347,86 @@ function categoryEditData () {
   sql='UPDATE `wallpaper_wallpaper` SET `category`="'+
   category+'" WHERE `category`="'+categoryManager.category+'"';
   $.post('php/update.php',{data:sql}, сategorySelectGetData);
+}
+function subCategoryManagerGetData () {
+  var category=$('select[name="subcategory-manager_select"]');
+  category=category[0].value;
+  var sql='SELECT * FROM `wallpaper_subcategory` WHERE `category`="'+category+'"';
+  $.post('../wallpaper/php/select.php',{data:sql},function (data){
+    var arr=JSON.parse(data);
+    subcategoryManagerForm(arr);
+  });
+}
+function subcategoryManagerForm (arr) {
+  var tr=$('.subcategory-manager_item');
+  $('.subcategory-manager_item:gt(0)').remove();
+  for (var i = 1; i < arr.length; i++) {
+    var trClone=tr[0].cloneNode(true);
+    var table=$('.subcategory-manager_table');
+    table[0].appendChild(trClone);
+  }
+  subcategoryManagerFormData(arr);
+}
+function subcategoryManagerFormData(arr) {
+  var idArr=$('.subcategory-manager_id');
+  var subcategoryArr=$('input[name="subcategory-manager_input"]');
+  var saveArr=$('button[name="subcategory-manager_save"]');
+  var delArr=$('button[name="subcategory-manager_delete"]');
+  for (var i = 0; i < arr.length; i++) {
+    $(idArr[i]).html(arr[i].id);
+    $(subcategoryArr[i]).prop('value',arr[i].subcategory);
+    $(subcategoryArr[i]).prop('disabled',true);
+    $(saveArr[i]).prop('value',arr[i].subcategory);
+    $(delArr[i]).prop('value',arr[i].subcategory);
+  }
+}
+function categoryFilterGetData () {
+  var sql='SELECT * FROM `wallpaper_category`';
+  $.post('../wallpaper/php/select.php',{data:sql},function(data){
+    var arr=JSON.parse(data);
+    var select= $('select[name="admin-category_filter"]');
+    var option = $(select[0]).children();
+    $(option).slice(1).remove();
+    for (var i = 0; i < arr.length; i++) {
+      var cloneOption=option[0].cloneNode(true);
+      select[0].appendChild(cloneOption);
+      cloneOption.innerHTML=arr[i].category;
+      cloneOption.value=arr[i].category;
+    }
+  });
+}
+function categorySelectGalery() {
+  subcategoryFilterGetData(this.value);
+}
+function subcategoryFilterGetData (category) {
+  var sql='SELECT * FROM `wallpaper_subcategory`'+
+  'WHERE `category`="'+category.trim()+'"';
+  $.post('../wallpaper/php/select.php',{data:sql},function(data){
+    var arr=JSON.parse(data);
+    var select= $('select[name="admin-subcategory_filter"]');
+    var option = $(select[0]).children();
+    $(option).slice(1).remove();
+    for (var i = 0; i < arr.length; i++) {
+      var cloneOption=option[0].cloneNode(true);
+      select[0].appendChild(cloneOption);
+      cloneOption.innerHTML=arr[i].subcategory;
+      cloneOption.value=arr[i].subcategory;
+    }
+    filterGalery();
+  });
+}
+function filterGalery () {
+  var sql='SELECT * FROM `wallpaper_wallpaper`';
+  var category=$('select[name="admin-category_filter"]').prop('value');
+  var subcategory=$('select[name="admin-subcategory_filter"]').prop('value');
+  console.log(category,subcategory);
+  if (category!='all'){
+    sql=    'SELECT * FROM `wallpaper_wallpaper`'+
+    'WHERE `category`="'+category+'"';
+  }
+  if (subcategory!='all'){
+    sql=    'SELECT * FROM `wallpaper_wallpaper`'+
+    'WHERE `category`="'+category+'" AND `subcategory`="'+subcategory+'"';
+  }
+  getGaleryData(sql);
 }
